ocr-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pytesseract==0.3.13
python-multipart==0.0.9
PyMuPDF==1.24.9
Pillow==10.4.0
numpy==1.26.4
